20221208 #1
- 감정분석 데이터 생성
- 참석자별 막대그래프 생성
- 참석자별 집중도 생성
- 참석자 리스팅 순서 호스트 최우선, 이름순으로 변경

+++ src/main/java/com/eura/web/model/DTO/GraphMidVO.java
@@ -8,8 +8,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class GraphMidVO {
-    List<Double> goodList;
-    List<Double> badList;
+    List<Integer> goodList;
+    List<Integer> badList;
     List<Integer> TotcalCntList;
     List<String> lvlList;
 }
